Migrate cleanup_datasets script to work with new db schema.

+++ scripts/cleanup_datasets/cleanup_datasets.py
@@ -2,6 +2,7 @@
 
 import sys, os, time, ConfigParser
 from datetime import datetime, timedelta
+from time import strftime
 from optparse import OptionParser
 
 new_path = [ os.path.join( os.getcwd(), "lib" ) ]
@@ -47,21 +48,22 @@ def main():
     h = app.model.History
     d = app.model.Dataset
     cutoff_time = datetime.utcnow() - timedelta( days=options.days )
+    now = strftime( "%Y-%m-%d %H:%M:%S" )
 
-    print "\n# Handling stuff older than %i days\n" %options.days
+    print "\n# %s - Handling stuff older than %i days\n" % ( now, options.days )
 
     if options.info_delete_userless_histories:
         info_delete_userless_histories( h, cutoff_time )
     elif options.delete_userless_histories:
-        delete_userless_histories( h, cutoff_time )
+        delete_userless_histories( h, d, cutoff_time )
     if options.info_purge_histories:
-        info_purge_histories( h, cutoff_time )
+        info_purge_histories( h, d, cutoff_time )
     elif options.purge_histories:
         if options.remove_from_disk:
             print "# Datasets will be removed from disk...\n"
         else:
             print "# Datasets will NOT be removed from disk...\n"
-        purge_histories( h, cutoff_time, options.remove_from_disk )
+        purge_histories( h, d, cutoff_time, options.remove_from_disk )
     elif options.info_purge_datasets:
         info_purge_datasets( d, cutoff_time )
     elif options.purge_datasets:
@@ -77,14 +79,16 @@ def info_delete_userless_histories( h, cutoff_time ):
     history_count = 0
     dataset_count = 0
     where = ( h.table.c.user_id==None ) & ( h.table.c.deleted=='f' ) & ( h.table.c.update_time < cutoff_time )
-    histories = h.query().filter( where ).options( eagerload( 'datasets' ) )
+    histories = h.query().filter( where ).options( eagerload( 'active_datasets' ) )
 
     print '# The following datasets and associated userless histories will be deleted'
     start = time.clock()
     for history in histories:
-        for dataset in history.datasets:
-            if not dataset.deleted:
-                print "dataset_%d" %dataset.id
+        for dataset_assoc in history.active_datasets:
+            if not dataset_assoc.deleted:
+                # This check is not necessary since 'active_datasets' are not
+                # deleted, but just being cautious
+                print "dataset_%d" %dataset_assoc.dataset_id
                 dataset_count += 1
         print "%d" % history.id
         history_count += 1
@@ -92,23 +96,31 @@ def info_delete_userless_histories( h, cutoff_time ):
     print "# %d histories ( including a total of %d datasets ) will be deleted\n" %( history_count, dataset_count )
     print "Elapsed time: ", stop - start, "\n"
 
-def delete_userless_histories( h, cutoff_time ):
+def delete_userless_histories( h, d, cutoff_time ):
     # Deletes userless histories whose update_time value is older than the cutoff_time.
     # The datasets associated with each history are also deleted.  Nothing is removed from disk.
     history_count = 0
     dataset_count = 0
-    where = ( h.table.c.user_id==None ) & ( h.table.c.deleted=='f' ) & ( h.table.c.update_time < cutoff_time )
+    h_where = ( h.table.c.user_id==None ) & ( h.table.c.deleted=='f' ) & ( h.table.c.update_time < cutoff_time )
 
     print '# The following datasets and associated userless histories have been deleted'
     start = time.clock()
-    histories = h.query().filter( where ).options( eagerload( 'datasets' ) )
+    histories = h.query().filter( h_where ).options( eagerload( 'active_datasets' ) )
     for history in histories:
-        for dataset in history.datasets:
-            if not dataset.deleted:
-                dataset.deleted = True
-                dataset.clear_associated_files()
-                dataset.flush()
-                print "dataset_%d" %dataset.id
+        for dataset_assoc in history.active_datasets:
+            if not dataset_assoc.deleted:
+                # Mark all datasets as deleted
+                d_where = ( d.table.c.id==dataset_assoc.dataset_id )
+                datasets = d.query().filter( d_where )
+                for dataset in datasets:
+                    if not dataset.deleted:
+                        dataset.deleted = True
+                        dataset.flush()
+                # Mark the history_dataset_association as deleted
+                dataset_assoc.deleted = True
+                dataset_assoc.clear_associated_files()
+                dataset_assoc.flush()
+                print "dataset_%d" % dataset_assoc.dataset_id
                 dataset_count += 1
         history.deleted = True
         history.flush()
@@ -118,32 +130,37 @@ def delete_userless_histories( h, cutoff_time ):
     print "# Deleted %d histories ( including a total of %d datasets )\n" %( history_count, dataset_count )
     print "Elapsed time: ", stop - start, "\n"
     
-def info_purge_histories( h, cutoff_time ):
+def info_purge_histories( h, d, cutoff_time ):
     # Provide info about the histories and datasets that will be affected if the purge_histories function is executed.
     history_count = 0
     dataset_count = 0
     disk_space = 0
-    where = ( h.table.c.deleted=='t' ) & ( h.table.c.purged=='f' ) & ( h.table.c.update_time < cutoff_time )
+    h_where = ( h.table.c.deleted=='t' ) & ( h.table.c.purged=='f' ) & ( h.table.c.update_time < cutoff_time )
 
     print '# The following datasets and associated deleted histories will be purged'
     start = time.clock()
-    histories = h.query().filter( where ).options( eagerload( 'datasets' ) )   
+    histories = h.query().filter( h_where ).options( eagerload( 'datasets' ) )   
     for history in histories:
-        for dataset in history.datasets:
-            if not dataset.purged:
-                print "%s" % dataset.file_name
-                dataset_count += 1
-                try:
-                    disk_space += dataset.file_size
-                except:
-                    pass
+        for dataset_assoc in history.datasets:
+            # Datasets can only be purged if their HistoryDatasetAssociation has been deleted.
+            if dataset_assoc.deleted:
+                d_where = ( d.table.c.id==dataset_assoc.dataset_id )
+                datasets = d.query().filter( d_where )
+                for dataset in datasets:
+                    if dataset.purgable and not dataset.purged:
+                        print "%s" % dataset.file_name
+                        dataset_count += 1
+                        try:
+                            disk_space += dataset.file_size
+                        except:
+                            pass
         print "%d" % history.id
         history_count += 1
     stop = time.clock()
     print '# %d histories ( including a total of %d datasets ) will be purged.  Freed disk space: ' %( history_count, dataset_count ), disk_space, '\n'
     print "Elapsed time: ", stop - start, "\n"
 
-def purge_histories( h, cutoff_time, remove_from_disk ):
+def purge_histories( h, d, cutoff_time, remove_from_disk ):
     # Purges deleted histories whose update_time is older than the cutoff_time.
     # The datasets associated with each history are also purged.
     history_count = 0
@@ -151,37 +168,39 @@ def purge_histories( h, cutoff_time, remove_from_disk ):
     disk_space = 0
     file_size = 0
     errors = False
-    where = ( h.table.c.deleted=='t' ) & ( h.table.c.purged=='f' ) & ( h.table.c.update_time < cutoff_time )
+    h_where = ( h.table.c.deleted=='t' ) & ( h.table.c.purged=='f' ) & ( h.table.c.update_time < cutoff_time )
 
     print '# The following datasets and associated deleted histories have been purged'
     start = time.clock()
-    histories = h.query().filter( where ).options( eagerload( 'datasets' ) )      
+    histories = h.query().filter( h_where ).options( eagerload( 'datasets' ) )      
     for history in histories:
         errors = False
-        for dataset in history.datasets:
-            if not dataset.purged:
-                file_size = dataset.file_size
-                dataset.deleted = True
-                dataset.file_size = 0
-                if remove_from_disk:
-                    dataset.flush()
-                    errmsg = purge_dataset( dataset )
-                    if errmsg:
-                        errors = True
-                        if errmsg.startswith( '# Error' ):
-                            # Only log errors, no need to log "shared history" dataset messages
-                            print errmsg
-                    else:
-                        print "%s" % dataset.file_name
-                else:
-                    dataset.purged = True
-                    dataset.flush()
-                    print "%s" % dataset.file_name
-                dataset_count += 1
-                try:
-                    disk_space += file_size
-                except:
-                    pass
+        for dataset_assoc in history.datasets:
+            if dataset_assoc.deleted:
+                d_where = ( d.table.c.id==dataset_assoc.dataset_id )
+                datasets = d.query().filter( d_where )
+                for dataset in datasets:
+                    if dataset.purgable and not dataset.purged:
+                        file_size = dataset.file_size
+                        dataset.deleted = True
+                        dataset.file_size = 0
+                        if remove_from_disk:
+                            dataset.flush()
+                            errmsg = purge_dataset( dataset )
+                            if errmsg:
+                                errors = True
+                                print errmsg
+                            else:
+                                print "%s" % dataset.file_name
+                        else:
+                            dataset.purged = True
+                            dataset.flush()
+                            print "%s" % dataset.file_name
+                        dataset_count += 1
+                        try:
+                            disk_space += file_size
+                        except:
+                            pass
         if not errors:
             history.purged = True
             history.flush()
@@ -195,7 +214,7 @@ def info_purge_datasets( d, cutoff_time ):
     # Provide info about the datasets that will be affected if the purge_datasets function is executed.
     dataset_count = 0
     disk_space = 0
-    where = ( d.table.c.deleted=='t' ) & ( d.table.c.purged=='f' ) & ( d.table.c.update_time < cutoff_time )
+    where = ( d.table.c.deleted=='t' ) & ( d.table.c.purgable=='t' ) & ( d.table.c.purged=='f' ) & ( d.table.c.update_time < cutoff_time )
 
     print '# The following deleted datasets will be purged'    
     start = time.clock()
@@ -217,7 +236,7 @@ def purge_datasets( d, cutoff_time, remove_from_disk ):
     dataset_count = 0
     disk_space = 0
     file_size = 0
-    where = ( d.table.c.deleted=='t' ) & ( d.table.c.purged=='f' ) & ( d.table.c.update_time < cutoff_time )
+    where = ( d.table.c.deleted=='t' ) & ( d.table.c.purgable=='t' ) & ( d.table.c.purged=='f' ) & ( d.table.c.update_time < cutoff_time )
 
     print '# The following deleted datasets have been purged'
     start = time.clock()
@@ -227,9 +246,7 @@ def purge_datasets( d, cutoff_time, remove_from_disk ):
         if remove_from_disk:
             errmsg = purge_dataset( dataset )
             if errmsg:
-                if errmsg.startswith( '# Error' ):
-                    # Only log errors, no need to log "shared history" dataset messages
-                    print errmsg
+               print errmsg
             else:
                 dataset_count += 1
                 print "%s" % dataset.file_name
@@ -251,42 +268,33 @@ def purge_datasets( d, cutoff_time, remove_from_disk ):
 
 def purge_dataset( dataset ):
     # Removes the file from disk and updates the database accordingly.
-    if dataset.dataset_file is None or not dataset.dataset_file.readonly:
-        # Check to see if another dataset is using this file.  This happens when a user shares 
-        # their history with another user.  In this case, a new record is created in the dataset
-        # table for each dataset, but the dataset records point to the same data file on disk.  So
-        # if 1 of the 2 users deletes the dataset from their history but the other doesn't, we need
-        # to keep the dataset on disk for the 2nd user.
-        if dataset.dataset_file:
-            for data in dataset.select_by( purged=False, filename_id=dataset.dataset_file.id ):
-                if data.id != dataset.id:
-                    return "# Dataset for deletion ( id %s ) points to a file on disk being shared by another user's history ( dataset id %s )\n" %( str( dataset.id ), str( data.id ) )
-        elif dataset.deleted:
-            # Remove files from disk and update the database
-            purgable = False
-            try:
-                dataset.purged = True
-                dataset.file_size = 0
-                dataset.clear_associated_files( purge = True )
-                dataset.flush()
-                if dataset.dataset.purgable:
-                    for shared_data in dataset.dataset.history_associations:
-                        if not shared_data.purged:
-                            break #only purge when not shared
-                    else:
-                        os.unlink( dataset.file_name )
-                        purgable = True
-            except Exception, exc:
-                return "# Error, exception: %s caught attempting to purge %s\n" %( str( exc ), dataset.file_name )
-            try:
-                if purgable:
-                    os.unlink( dataset.extra_files_path )
-            except:
-                pass
-        else:
-            return "# Error: '%s' has not previously been deleted, so it cannot be purged\n" %dataset.file_name
+    if dataset.deleted:
+        # Remove files from disk and update the database
+        purgable = False
+        try:
+            dataset.purged = True
+            dataset.file_size = 0
+            dataset.flush()
+            for shared_data in dataset.history_associations:
+                # Check to see if another dataset is using this file.  This happens when a user shares 
+                # their history with another user.  In this case, a new record is created in the dataset
+                # table for each dataset, but the dataset records point to the same data file on disk.  So
+                # if 1 of the 2 users deletes the dataset from their history but the other doesn't, we need
+                # to keep the dataset on disk for the 2nd user.
+                if not shared_data.deleted:
+                    break #only purge when not shared
+            else:
+                os.unlink( dataset.file_name )
+                purgable = True
+        except Exception, exc:
+            return "# Error, exception: %s caught attempting to purge %s\n" %( str( exc ), dataset.file_name )
+        try:
+            if purgable:
+                os.unlink( dataset.extra_files_path )
+        except:
+            pass
     else:
-        return "# Error: '%s' has dependencies, so it cannot be purged\n" %dataset.file_name
+        return "# Error: '%s' has not previously been deleted, so it cannot be purged\n" %dataset.file_name
     return ""
 
 class CleanupDatasetsApplication( object ):
